Eval

Description

This module implements expression evaluation.

Documentation

type Eval a = ReaderT ValueEnv (Except TypeError) a

The expression evaluation monad.

type ValueEnv = Map Iden Value

A binding of identifier to values, in which we can evaluate expressions. Note that we use a lazy map to allow recursive bindings.

data Value

The result of evaluating an expression.

Constructors

LitV Lit
Literal value.

LambdaV Iden Expr ValueEnv
Closure.

BuiltinV Builtin
Built-in function.

data Builtin

A built-in function.

Currently, the following functions are available:

Unary double operations: neg

Binary double operations: add, mul

Predicates: eq, gt, lt

Conditionals: if

bind :: Binding -> ValueEnv -> Either TypeError ValueEnv

eval :: ValueEnv -> Expr -> Either TypeError Value

Runs an Eval action.