HindleyMilner

Description

This module exposes a higher-level API to run computations in our toy implementation of the Hindley-Milner type system. It also provides various common built-in functions via defaultEnv.

The underlying implementation is split into three modules:

HindleyMilner.Parse, responsible for parsing text into expressions

HindleyMilner.Infer, responsible for inferring the type of each expression

HindleyMilner.Eval, responsible for evaluating well-typed expressions. This module also provides utilities for extending the implementation with new built-ins.

Interpreter

interpret :: MonadState Env m => Text -> m Result

Interpret a command.

For example, the following will compute the value six:

result = flip evalState defaultEnv $ do
  interpret "x = 1"
  interpret "y = add x 2" -- Assign y = x + 2 = 3.
  interpret "mul x y" -- Return x * y = 6.

data Result

The result of interpreting a command.

Constructors

Computation Value
A pure value.

BoundIdentifier Iden Type
A binding was added to the environment.

InvalidSyntax ParsingError
The command could not be parsed.

InvalidType TypeError
The command was parsed but did not have a valid type.

Environment

data Env

An evaluation environment.

defaultEnv :: Env

A default environment with bindings for common functionality:

Numeric operations: add, mul, neg, ...

Conditional and predicates: if, eq, gt, lt, ...

envLookup :: Iden -> Env -> Maybe (Type, Value)

Looks up an identifier in the environment.

Types

data Type

The type of an expression.

Constructors

VarT TypeVar
A type variable.

ConstT Text
A type constant (e.g. doubles or strings).

ArrowT Type Type infixr 5
A function type.

data TypeVar

A type variable.

displayType :: Type -> Text

Pretty-prints a type. All type variables are represented as $N, numbered starting from 1 in the order they are displayed left to right (so $1, $2, ...).

Values

data Value

The result of evaluating an expression.

Constructors

LitV Lit
A literal value.

ClosureV (Value -> Value)
A closure.

UndefinedV
Something went wrong.

data Lit

A literal value.

Constructors

DoubleL Double

StringL Text

BoolL Bool

displayLit :: Lit -> Text

Errors

type ParsingError = ParseErrorBundle Text Void

Error thrown during parsing.

displayParsingError :: ParsingError -> Text

Pretty-prints a parsing error.

data TypeError

Error thrown when an expression is ill-typed.

Constructors

InfiniteType Type Type
Failed to unify a type which occurs in an arrow type.

UnificationFailure Type Type
Failed to match two types.

UnboundVariable Iden
Encountered an undefined identifier.

displayTypeError :: TypeError -> Text

Pretty-prints a type error.