HindleyMilner.Eval

Description

This module implements expression evaluation.

Environment

data Env

An evaluation environment.

envLookup :: Iden -> Env -> Maybe (Type, Value)

Looks up an identifier in the environment.

Computations

eval :: Expr -> Env -> Either TypeError (Type, Value)

Evaluates an expression in the given environment.

data Value

The result of evaluating an expression.

Constructors

LitV Lit
A literal value.

ClosureV (Value -> Value)
A closure.

UndefinedV
Something went wrong.

Extending environments

Safely

bind :: Binding -> Env -> Either TypeError Env

Adds a binding to an existing environment.

Unsafely

unsafeEnv :: Iden -> (Type, Value) -> Env

Generates an environment with a single identifier bound. The type and value must match, no checks are performed.

This is useful in combination to add "built-in" functionality. For example, to generate an environment with a sin operation, we could do something similar to:

env = unsafeEnv "sin" (tp, ClosureV fn) where
  tp = doubleType -> doubleType
  fn (LitV (DoubleL d)) = LitV $ DoubleL $ sin d
  fn _ = UndefinedV