HindleyMilner.Infer

Description

This module implements expression typing. It is the second step in the interpretation process (after parsing in Parse and before evaluation in Eval).

Type inference is done in two phases:

Discovery, where we walk the AST, assign fresh type variables where appropriate, and emit constraints.

Unification, where we unify all types using the emitted constraints and output the resulting expression's type.

Note that let bindings are generalized:

let f x = x in if (f True) (f 1) 0 -- OK.
let g f = if (f True) (f 1) 0 in let f x = x in g f -- Not OK ("generalized" once in @g f@).

Types

data Type

The type of an expression.

Constructors

VarT TypeVar
A type variable.

ConstT Text
A type constant (e.g. doubles or strings).

ArrowT Type Type infixr 5
A function type.

displayType :: Type -> Text

Pretty-prints a type. All type variables are represented as $N, numbered starting from 1 in the order they are displayed left to right (so $1, $2, ...).

newtype TypeVar

A type variable.

Constructors

TypeVar

Fields

unVar :: Int

boolType :: Type

doubleType :: Type

stringType :: Type

Inference

infer :: Expr -> TypeEnv -> Either TypeError Type

Infer an expression's type.

data TypeEnv

All known type signatures. We store signatures rather than types to be able to generalize them appropriately at usage sites (i.e. generalize let bindings but keep function arguments as-is).

boundTo :: Type -> Iden -> TypeEnv

Creates a singleton TypeEnv for the given type. The binding is as polymorphic as possible.

lookupType :: Iden -> TypeEnv -> Maybe Type

Extracts a type from the environment.

data TypeError

Error thrown when an expression is ill-typed.

Constructors

InfiniteType Type Type
Failed to unify a type which occurs in an arrow type.

UnificationFailure Type Type
Failed to match two types.

UnboundVariable Iden
Encountered an undefined identifier.

displayTypeError :: TypeError -> Text

Pretty-prints a type error.