HindleyMilner.Parse

Description

This module implements parsers for our language. It is the first step in the interpretation process (followed by type inference in Infer, and finally evaluation in Eval). More specifically, this module is responsible for transforming Text into expressions (Expr).

Examples:

Factorial 10

let fact n = if (eq n 0) 1 (mul n (fact (add n (neg 1)))) in fact 10

Expression types

data Expr

An expression.

Constructors

VarE Iden
A variable (e.g. abc).

LitE Lit
A literal value.

AppE Expr Expr
An application.

LambdaE Iden Expr
A lambda.

LetE Binding Expr
A let binding (e.g. let f x = add x 2 in f 1).

type Iden = Text

An identifier (e.g. variable name).

data Lit

A literal value.

Constructors

DoubleL Double

StringL Text

BoolL Bool

displayLit :: Lit -> Text

data Binding

A bound identifier.

Constructors

Binding

Fields

bindingIden :: Iden

Bound reference.

bindingArgs :: [Iden]

Arguments (if the binding is a lambda).

bindingExpr :: Expr

Bound expression.

Parsing

type Parser a = Parsec Void Text a

A parser specialized to our implementation.

exprParser :: Parser Expr

Returns an expression parser.

Five types of expressions are supported:

Identifiers: IDEN, must start with a lower-case letter then alpha-numeric characters

Literals: LIT (numbers, True, False)

Function applications: FN_IDEN ARG_IDEN

Bindings: let BINDING in EXPR

Groups: (EXPR)

bindingParser :: Parser Binding

Returns a parser for bindings (the subsection between let and in). This parser is exported to allow extending the built-in environment.

The syntax is REF_IDEN [ARG_IDEN...] = BODY_EXPR where REF_IDEN is the identifier of the binding, ARG_IDEN is the identifier of an argument (there can be 0 or more), and EXPR_BODY is the body of the binding.

type ParsingError = ParseErrorBundle Text Void

Error thrown during parsing.

displayParsingError :: ParsingError -> Text

Pretty-prints a parsing error.