Infer

Description

This module implements expression typing. It is the second step in the interpretation process (after parsing in Parse and before evaluation in Eval).

Type inference is done in two phases:

First, we walk the AST, assign fresh type variables where appropriate, and emit constraints.

Then, we unify all types using the emitted constraints and output the resulting expression's type.

Note that let bindings are generalized:

let f x = x in if (f True) (f 1) 0 -- OK.
let g f = if (f True) (f 1) 0 in let f x = x in g f -- Not OK ("generalized" once in @g f@).

Types

data Type

The type of an expression.

Constructors

VarT Var
A type variable.

ConstT Text
A type constant (e.g. doubles or strings).

ArrowT Type Type
A function type.

displayType :: Type -> Text

Pretty-prints a type. All type variables are represented as $N, numbered starting from 1 in the order they are displayed left to right (so $1, $2, ...).

newtype Var

A type variable.

Constructors

Var

Fields

unVar :: Int

boolType :: Type

doubleType :: Type

stringType :: Type

Inference

infer :: TypeEnv -> Expr -> Either TypeError Type

Infer an expression's type.

type TypeEnv = Map Iden Sig

All known type signatures. We store signatures rather than types to be able to generalize them appropriately at usage sites (i.e. generalize let bindings but keep function arguments as-is).

data Sig

A closed type signature.

Constructors

ForAll [Var] Type

data TypeError

Error thrown when an expression is ill-typed.

Constructors

InfiniteType Type Type
Failed to unify a type which occurs in an arrow type.

TypeMismatch Type Type
Failed to match two types.

UnboundIden Iden
Encountered an undefined identifier.